Password syncing
This commit adds a job which is intended to run on a clean jenkins
deployment to sync passwords from the safe to jenkins' credential
store.  The list of passwords to sync is hard-coded in
pull_passwords.py and will need to be pulled out at a later date.

At time of writing add_jenkins_cred.groovy supports text or file
credential types.

As this is a bootstrap job, we have tried to limit the amount of
credentials that already need to be in place before this runs.  The
only one that will need to be manually created are
service_account_jenkins_ssh_key and rpc-jenkins-svc-github-key.
The former is the ssh key assigned to the service account jenkins user
(in jenkins itself) which gets used when running jenkins-cli.jar, and
the latter is the GitHub private key to install pwsafe.

Connects https://github.com/rcbops/u-suk-dev/issues/801