Patience required when waiting for quota
Jobs should not fail when there is not enough quota, however most jobs
are long running, so if theres no quota on the first try, its unlikely
that the situation is different after 5x30s delay.

This commit sets 120x60s retries. After a couple of hours
there should be some quota - or there is a problem that requires human
intervention.