Remove swift from ceph job
Currently, all our builds include swift even when ceph is being used.
This commit explicitly removes swift from being deployed for the ceph
job so that we have at least one job where swift is not being used.
This is necessary to partially cover the use case of swift not being
present in an environment.