Calculate and set ANSIBLE_FORKS
This commit adds a function to common.groovy to calculate a reasonable #
for ANSIBLE_FORKS, and then modifies any functions that call deploy.sh
or openstack-ansible to pass in the env var ANSIBLE_FORKS.

The reason for doing this is that on rpc-openstack's liberty and mitaka
branches, deploy.sh does not correctly set forks when running
rpc-openstack playbooks.  This is resulting in a number of job failures
due to ssh connection failures because it's picking up the forks = 15
in rpcd/playbooks/ansible.cfg.

We also set env var ANSIBLE_SSH_RETRIES=3 when running deploy.sh or
openstack-ansible.  This env var is ansible 2.x and above only, so will
have no affect on liberty and mitaka branches.

Connects https://github.com/rcbops/u-suk-dev/issues/1406