Update deploy_sh to set FORKS env var
This commit updates deploy_sh function to set the FORKS env var to the
same value as ANSIBLE_FORKS.  We also update multi-node-aio.yml to
remove passing in FORKS=10 to deploy_sh.

Note that calc_ansible_forks will return 10 for an OnMetal instance as
they seem to have > 10 CPUs and calc_ansible_forks will return 10 when
there are > 10.  As such, this commit should keep the behaviour the
same for the OnMetal job.