Reduce slave executors to 2
Currently most slaves are short-lived, so they
do not need more than 1 executor. Having 15
causes the Jenkins dashboard to be rather long
to scroll down and my finger occasionally gets
a cramp. I'd rather avoid that.

So, for a better world, we're reducing the
executors for any slave nodes created to 2.