Add post merge trigger for jjb job
This ensures that jenkins-jobs is run after merges to rpc-gating.

An intermediate job is used to call the actual JJB job. This is to work
around a problem of parameter defaults not being correctly expanded by
the git scm plugin when polling for changes after recieving a push event
from github.