Create defaults file
This PR moves a bunch of job defaults to global defaults.

There is a huge net benefit of doing this, however it does mean we can
remove the raft of default variables set in RPC-AIO.yml and
RPC-Artifact-Build.yml and should we need to change REGION quickly, for
example, we can do it on place versus 3.