Set overrides to fix OnMetal tempest failures
This commit sets some openstack-ansible ansible overrides to fix a
number of tempest failures we see on the OnMetal jobs.  To summarise:

- use qemu instead of kvm (running nested virt (kvm) seems unstable and
  we end up getting a number of job failures that we do not see on the
  single node AIO or hardware labs)
- set reseller_admin_role in tempest.conf to match what gets created by
  swift (this is only an issue on stable/mitaka, the default value is
  already correct on later branches)
- disable personality in tempest.conf as our nova service by default
  disallows file injection (again, only needed on stable/mitaka as
  later versions of tempest seem to detect this correctly)

Ideally, we'll want to find the root cause of the kvm failures and try
to work around this so that we can continue to use kvm, but for now
this unblocks the gate and allows us to make use of the OnMetal build.