Skip teardown if meta key set
Currently, if you add the Cleanup stage to a build there is no way
clean way of preventing the teardown should you end up needing to
preserve the instance for investigation.

With this commit, you can now skip the teardown of an instance by
running the following mid-way through the build:

```
nova meta <instance> set preserve=yes
```

or

```
nova meta <instance> set preserve=true
```

For this to work, we've had to move the slave teardown to a separate
playbook so we can check for the existence of an ansible fact.  We also
create an ansible.cfg to write facts to disk so we can use these facts
in destroy-jenkins-slave.yml.  A new playbook called write_facts.yml is
added so we explicitly gather facts and write to disk.  We could
probably shimmy these into another playbook, but keeping it separate
will make it easier when we start to deploy to non-cloud instances.

Lastly, we update setup-jenkins-slave.yml by removing a bunch of defunct
tasks which never get executed as the playbook is called with a --limit
that does not include localhost.