Artifact job refactor
This patch aims to refactor the Artifact jobs so
that they more effectively use stages, and share
a common structure. The detailed steps are all in
the groovy file, whereas the pipeline logic is
formed into stages in the job dsl.

An additional job is also added to build apt
artifacts. This job runs on a jenkins slave rather
than a single-use node.

A common script to get the credentials into the
environment is used for all jobs.