Allow formatted Stages/Steps to be unconditional
The conditionalStage/conditionalStep functions appear
to largely be used to consistently format the output
when executing them, rather than to actually make them
conditional.

This is evidenced by the fact that the conditional
stages when executing on the public cloud instances
are absolutely not conditional and instead are required.

This patch renames the two functions, making the default
for each function to be to run the Stage/Step
unconditionally and providing a parameter to change it
for the stage.

It then also cleans up the unnecessary STAGES entries in
the Jenkins jobs as the stages are no longer optional.