Add git artifact job and job control vars
In order to update the git artifacts, another job is added
to the pipeline.

The three primary control points (pull, replace, push) for
all artifact jobs are added to allow fine grained control
of how the jobs behave and to provide the option of replacing
existing artifacts which may be needed from time to time.

Some adjustments are made to the job definition to make it
a little easier to read.