Use the 14.04.4 Trusty image for artifacting jobs
In order to be able to use apt artifacts successfully
to build the same tag in the same way today and next
year we need to fix the starting point to ensure that
the host OS does not have packages installed which are
later than the packages in the artifacted apt repo.

A 14.04.4 image has been prepared for Trusty and serves
as that starting point.

This is currently only implemented for the artifact
builds, but will eventually be used for all builds.