Configure apt artifact source when necessary
When a more restricted setup is used to ensure
that an artifacted deployment only uses the
prepared artifacts, the jobs fail as the jenkins
slave cannot install the JRE which comes from
the ubuntu updates repo.

This patch ensures that the slave is prepared
with the apt artifact repo source prior to
trying to install anything.