Temporarily disable cinder-backup from being tested
The cinder-backup service is usable on an AIO deployment, but
deploying on a multi-node deployment seems to result in the service not
working.  I'm not a storage expert, but from looking briefly it appears
that the open-iscsi package is missing on the storage nodes, which
prevents the service from backing up a volume.  The package open-iscsi
gets installed by nova on the compute nodes, and since an AIO deploys
nova-compute, cinder-volume, and cinder-backup to the host itself this
explains why cinder-backup is functional there.

I will create an upstream bug and fix for this but in the mean time
it's probably best to just disable this from being tested.

Connects https://github.com/rcbops/u-suk-dev/issues/1643