Re-enable cinder-backup on RPC-AIO jobs
The RPC-AIO jobs are currently failing tempest runs because it's
attempting to test cinder-backup while cinder backup is not being
deployed.  This commit keeps cinder-backup disabled for the OnMetal
AIOs but re-enables it for the RPC-AIOs.

Side note, the OSA AIO boostrap sets tempest_volume_backup_enabled
to true, which is why disabling this service did not affect the
OnMetal AIOs.