Switch artifact job branch and update base images
The artifacts-14.0 branch has become dated and will be removed.
Its content was moved into a new branch in preparation for a PR
to rpc-openstack/master and some new fixes have gone into it
since then. The fixes in there haven't been applied into the
artifacts-14.0 branch and the periodics on that branch are now
interfering with the artifacts published by manual job execution
on the artifacts branch. To resolve this the jobs should be
switched over to the artifacts branch.

Also, now that we have two base images configured in the same
way for both Trusty and Xenial, switch to using them for the
jobs.