UG-551 Use OSA SHA defined by RPCO with multi-node
This change ensures the version of openstack-ansible being tested on
multi-node AIOs is the one defined by the submodule in rpc-openstack.
To override the OSA SHA, add a new commit to RPCO and then test that
version of RPCO.

This change removes the ability to specify the OSA SHA as a job
parameter. Affected jobs will now use the SHA specified by the version
of RPCO being tested instead of defaulting to the HEAD of the relevant
OSA branch, this should remove the chance of incorrect test results due
to a mismatch between what RPCO specifies the OSA SHA should be and what
Jenkins uses. This change also helps to enforce the principle that
configuration should be in code.