Process apt artifact stage before using node
The node used for apt artifact builds only has one
executor. The job currently requires access to the
node before determining whether the stage is
necessary or not, resulting in any other jobs in
the pipeline waiting for a free executor.

This moves the stage check ahead of the node use
so that the node never needs to be used if the
stage is not included.