UG-605 Fix script not found in periodic cleanup
Periodic cleanup has the following rough structure:
node(){
 clone rpc-gating
 ...
 docker stuffs
 ...
 cleanup shell scripts
}

The problem is that the workspace used after "docker stuffs" is not the
same workspace that is used for the initial clone. That means that the
cleanup shell scripts can't be found. As a workaround, this commit
moves the cleanup shell scripts to before the docker step.