Update artifact/AIO jobs for newton-14.1
This adds the artifact build tests to the
master and newton-14.1 branches and cleans
up the references to the old artifacts
branch.

The AIO image is set to one that will never
change to ensure that the apt artifacts are
always newer than the existing packages on
the host which is being used.