Change AIO job axes for an increased test matrix
In order to cater for a more complex test matrix,
the AIO job axes have been updated as follows.

series:
The branch information to inform the git checkout
executed by jenkins, eg: master.

image:
The image to execute the job on, eg: trusty.

action:
The action to execute, eg: deploy, minor upgrade,
major upgrade, leapfrog.

scenario:
The scenario to configure when executing the job,
eg: swift, ceph.

In order to prepare for the master branch
becoming Ocata, the trusty builds are excluded
for the branch.