Add leapfrog AIO jobs
In order to test the leapfrog upgrades from
kilo to newton, jobs have been added to allow
the deployment of kilo and the leapfrog
upgrade to the newton-14.1 branch.

The job for the kilo deployment is to test
any new PR's to Kilo, if need be, and also
to validate via a periodic job that Kilo can
actually deploy. This will help to differentiate
between failures due to a kilo deployment, or
failures due to the leapfrog upgrade process.

The stages in the leapupgrade jobs are
deliberately reduced to a minimum set. This
is a starting target - once the job is showing
success more stages will be added to find
other potential bugs.