UG-618 Add MaaS entity removal to periodic cleanup
While entities tied to cloud instances are removed when those instances
are deleted, entities that are manually created are not. For instance
entities created for multi node aio VMs may not be cleaned up if a.
job is aborted.

This commit adds entity cleanup to the public cloud step of the
periodic cleanup job.

This commit also:
 * Adds agent token cleanup
 * Add multi region cloud server cleanup
 * Add logging to periodic cleanup
 * Changes lint.sh pre-commit hook from hard to symbolic.
   This will solve problems with lint when switching branches.