Remove rpc_release from test_holland
The value for rpc_release is set via group_vars for
all versions of RPC-O. The playbook should inherit
it from there instead of trying to override it.

This has been working so far due to the rpc_release
override set in user_*.yml files, but as we try to
cut down on those problems like this are likely to
arise.