UG-625 Adjust Docker cleanup
Removes images that are not the latest ubuntu that are more
than a day old.

This change is to resolve a found race condition in which the
cleanup job could run at a time and remove a newly created image
before the job it was created for ran the image.