UG-618 Add exception handler to webhooks.groovy
This will allow test jobs to cleanup after themselves even if an
exception is thrown. Note that by default the cleanup stages are not run
as this job is intended to create a long running node.

This is a separate commit so that changes to isolate the addition
of the try block from other changes to the file.