Fix execution of tempest on AIO
AIO builds are currently failing with a "No such file or directory"
error when running tempest on the AIOs.  This appears to be a result of
running tempest unwrapped with quotes around the tempest command, which
bash does not like:

root@ranxp-16-8323:~# 'cd /tmp && ls -al'
-bash: cd /tmp && ls -al: No such file or directory
root@ranxp-16-8323:~#

This commit simply creates a temporary variable depending on whether we
are using a wrapper (and need to quote) or not.