UG-625 Docker Cleanup tweak
Updates docker script to remove images on double digit hours. This
is needed as the `docker images -a` does not switch to include
days at 24 hours.

The `egrep` is required as the version of docker on the build node
does not include the `--format` flag in the cli that would allow
use of the `.CreatedAt` template to better filter images to delete.