UG-607 Ensure pull requests link to Jira issues
Add a line of the form `Issue: [Jira key](https://link/to/key)` to the
description of pull requests in tracked repositories.

This change defines jobs to update pull requests for rpc-openstack and
rpc-gating. Developers only add an issue key reference to commit
messages. This job runs a script to update the pull request description
body with a link to the issue to simplify accessing it.