UG-618 Add MaaS-Jira Webhook Translator
* Add MaaS-Jira Webhook Translator

This is a small flask app that receives webhooks from MaaS notifications
and creates jira tickets via the jira API.

While the app is small unfortunately there is quite a lot of python
packaging and deployment code around it.

This commit reuses the nginx that is already deployed to the webhook
proxy node. Nginx receives http requests and passes them on to
webhook translator via a fastcgi socket. The webook translator daemon
is managed as a systemd service and started by socket activation.

Other notes:
 * Setuptools' console scripts function is used to create
   startup scripts that are installed into the deployment venv.
   This allows the daemon to run within a venv without systemd
   having to know how to activate a venv.
 * get_maas_token_and_url.py has been replaced with maasutils.
   This includes the original functionality and adds the ability
   to set the webhook token for a MaaS account. This token is sent
   with MaaS notifications and verfied by webhook translator to ensure
   that requests that didn't originate from MaaS don't create tickets.
 * Webhook translator has some unit tests and these are called from
   tox in lint.sh. Perhaps that should become run-tests.sh or similar in
   future.

* Add exception handler to webhooks.groovy

This will allow test jobs to cleanup after themselves even if an
exception is thrown. Note that by default the cleanup stages are not run
as this job is intended to create a long running node.

This is a separate commit so that changes to isolate the addition
of the try block from other changes to the file.

* Allow webhook proxy instance name to be set