Do docs check without resource allocation
Currently all the artifact jobs do the resource
allocation prior to checking whether the patch
is a docs-only check. This wastes resources
unnecessarily.

Also, the docs check is currently trying to
check the wrong place, resulting in the tests
not being executed.

This ensures that the docs check is done on the
CIT slave prior to processing any stages. It
also ensures that the check is executed in the
right folder.