UG-637 Check for git repo in docs check
It is important to check that a git repo has been supplied before
checking if all changes are doc related. Otherwise a misconfigured
job may supply a non repo dir, and the build will get skipped and
marked as successful rather than failing as expected.

This commit raises an exception if the directory supplied to
common.is_doc_update_pr is not a valid git repo.