Add dashboard link to MaaS Jira tickets
This change adds a link to the intelligence dashboard for an alert.
This will allow a user to see historical data related to the alert.

This change also adds a real sample MaaS webhook payload as reality
doesn't quite match the sample given in the docs.

Tests are also udpated to reflect the new sample payload.