Clean up all slaves
This commit addresses a problem where the Jenkins section of the
periodic cleanup job fails because its running on a slave that doesn't
have access to the Jenkins API.

When fixing this I realised that most of the cleanups are node
specific, so we actually need to run them against every node.
This job introduces a parallel stage to clean up various objects on
all slaves, then a second stage to run public cloud and Jenkins node
cleanups on a single internal slave.