Make the recheck trigger phrase more sensible
Currently the trigger phrase and the status
context do not match, making it very hard to
figure out the trigger phrase.

This patch makes them match so that it is
more intuitive.

'repo' is removed from the trigger phrase
as the web hooks will not accept triggers
from any repo other than the originating
one anyway.