AR-106 Do not pull apt data from rpc-repo
For the apt artifact creation, rpc-repo is
primarily a publishing point and a place to
store a backup of the database. We should
therefore not be pulling the data from it
every time we run a job, but instead just
execute the job on the long-running node
which has the current database.