UG-646 Skip aio config if handled in repo
This commit adds a check for a capabilities file which indicates
that the branch being tested handles its own configuration.
Once all branches contain this functionality, then the rpc-gating
functionality for configuring rpc-o can be completely removed.

This commit also ensures that the JJB axes are passed through to the
job's environment, so that in repo scripts can make configuration
choices based on the scenario being tested.