[LA-344] Include more tests in leapfrog
MaaS, Tempest, and Kibana should be tested after the leapfrog.

NOTE: Holland will not be tested as part of these builds.
When we'll consider adding the support role for inclusion
in newton-14.1 branch, we can add Holland to STAGES in periodics.
This also means we do not need to explicitly add Holland to STAGES
for leapfrog jobs.