UG-640 Updates Jira issue key
Updates get_jira_issue_key function to only compare changes from
commits provided from the PR using the upstream repo, rather than
the forks master branch (which may not be up to date).

Also changes to allow the issue key to be included in the body of
the commit rather than requiring it to be at the beginning of the
subject.