De-dupe STAGES param
This commit de-dupes the STAGES param, allowing us to tweak specific
STAGES depending on action and trigger.

More specifically, with this change in place we can gate [1], which
allows us to remove the deployment of the rpc support role on PR
builds.

Note that `DEPLOY_SUPPORT_ROLE` will only be available once [1]
merges, however setting this in this PR saves us the extra PR once
[1] lands.

[1] https://github.com/rcbops/rpc-openstack/pull/2343