UG-666 Updates tempest results location
Master/Ocata places tempests result in a different path than pre-
Ocata. This updates command to attempt to find the location of
the >=Ocata `tempest_results.xml` file if exists and includes that
in path of results files that are attempted to be copied.

Additionally removes `deploy_vm` code in tempest steps for changes
from #299