UG-646 Supply MaaS env vars to RPC scripts
Previously these vars were added to overrides via the rpc-gating
aio_config playbook. This configuration is now done in rpc-o, so we
need to ensure the token and url variables are available in the
environment for the rpc-o scripts to use.

This is the second attempt, first attempt:
  * introduction: https://github.com/rcbops/rpc-gating/pull/296
  * revert: https://github.com/rcbops/rpc-gating/pull/297

The first attempt failed because it attempted to run a script from
rpc-gating on a single use slave that didn't have a checkout of
rpc-gating. This version ensures that the the script is run on a
long-running slave that will have the appropriate checkout.