Update multi-node job to better match rpc_aio
There has been significant drift between the way `multi_node_aio.yml`
and `rpc_aio.yml` are defined. This commit seeks to improve the
consistency between the files to simplify the process of expanding the
jobs performed on MNAIOs.

This changes copies changes made by:
- 7862381573c7e6cff29e2aa6898c8c27a6b8d121
    - `STAGES` is separated into multiple components to prevent the need
      for duplication when additional jobs are added
- 245c8304d04e980af114480d1a8ccbd0a0473cfb
    -  the job axes have been adjusted to rename `context` to `image`
       and add `action`
    - trusty builds are excluded for master

The set of parameters `rpc_deploy_params` is included to remove
duplication, this has resulted in USER_VARS being adjusted to make use
of the separate components it defines as well as removing some of the
items from `environment_vars` now that they are specified as parameters.