La-366 Ensure kibana branches make sense
Most of the upgrade jobs use the target branch as the trigger branch,
for exmaple changes to newton-14.1 have an upgrade test from kilo
to ensure the changes don't break upgrades.

However the kilo leapfrog job tests changes to kilo with a leapfrog
to newton-14.1 This alternate dirction upgrade requires some variable
changes to succeed.

In this commit the kibana branch variables are set to sensible
values for the kilo-newton leap job, as the defaults do not work.
Mostly because UPGRADE_FROM_REF is not set for kilo.