RE-51 Force known pkg version in workspace venv
When we setup the venv, we set it up without
the host's versions of pip, setuptools and wheel
(if possible) so that we can independently set
the version we want.

In order to ensure that we do not try to upgrade
or downgrade pip/setuptools/wheel in an existing
venv, we remove it from the requirements.

We only force a re-installation of pip, setuptools
and wheel into the venv if the pip version is not
available or incorrect.

This patch does a revert of some of the changes
in 73ad014