Move node allocation to a common function
This allows us to add hooks such as workspace clearup
and venv preparation. This will prevent us from having
to prepare the venv multiple times.

Adds a vars --> pipeline_steps symlink. This allows
us to load the rpc-gating repo as global library in jenkins, and have
it implicitly supplied to every job.

The common prefix is removed from common.func calls within the common
module, as they are not valid when common is loaded as a library.

Cleans up the now unnecessary lines from each job:
* rpc-gating clones
* groovy load statements
* deletedir calls