RE-48 Balance instances across regions
This commit allows jobs to attempt to create instances in a list
of regions, then fallback to a secondary list of regions if
the primary set fails.

A result of this is that the region where an instance is created is not
known at the beginning of the job, so the region must be recorded in
order for clean to be executed against the correct region. The region
is recorded in the ansible inventory file that is generated during
the allocation step. That file is stashed and retrieved so that
cleanup can succeed even if its executed on a different slave to
the allocation step.