Add support for leapfrog upgrades
This commit replicates the work in rpc_aio.yml to bring leapfrog upgrade
support to the multi-node aio jobs.

The following config is added to `USER_VARS`:
  - lxc_container_vg_name: "USE_DIR_NOT_VG"
    In Kilo the role `lxc_container_create` attempts to build the
containers in a volume group if the specified one exists. The multi-node
setup creates a volume group with the same name that does not have
sufficient space. Overriding the var in this way causes the containers
to be directory backed.
  - glance_default_store: "swift"
    This should already be set but due to differences in how config is
managed in the releases currently it needs to be set manually for the
time being.
  - neutron_legacy_ha_tool_enabled: true
    The leapfrog upgrades script requires this to be set.

The user variables files names are not consistent between the releases,
the configuration preparation stage has been adjusted so that it does
not expect specific names.

To overcome regular SSH errors when deploying Kilo, the ssh_retry branch
of Ansible has been enabled. Due to the way multi-node AIOs are
bootstrapped, this cannot be done with environment variables and so the
node is re-bootstrapped wit the appropriate code.