UG-646 add library statement to every job
Currently the rpc-gating library is loaded implicitly, jenkins injects
it into the groovy environment before the script is started. The
problem with this approach is that the verison of rpc-gating used
cannot be overridden.

This version uses the RPC_GATING_BRANCH parameter to set the
version of rpc-gating that is loaded. This will return our ability
to test other jobs (such as RPC AIOs) against changes to gating.

This approach also makes it explicit in each job where the library
is coming from. Once this merges, we'll turn off implicit loading
in Jenkins global config.