RE-106 Specify leapfrog upgrades environment_vars
Define the variable `environment_vars` to prevent leapfrog upgrades
failing due to its absence.

A recent change appears to have broken leapfrog upgrades on multi-node
AIOs which was not detected due to the inherent race that exists with
the current testing and merging process. `environment_vars` was no
longer defined when `deploy.upgrade_leapfrog` is run causing a failure.
In reality this failure has exposed a hidden bug because
`environment_vars` should not be defined as a by-product of some other
part of the code and so this change should now address the underlying
issue.