UG-646 Removing RPC_GATING_REPO param
The Jenkins shared library mechanism requires a fixed repo to load
libraries from. This parameter implies that the rpc-gating library
can be loaded from any repo, so should be removed for clarity.

New versions of the library can be tested by pushing branches to
rcbops/rpc-gating and loading from there.