RE-103 Store rpc-gating venvs
RPC-Gating currently builds a venv in the workspace of every node
that is used as part of a job. This is useful as we can rely on the
right version of various tools to be available, however it is network
intensive and slow. This commit reduces the time taken by using a pre
built venv when one is available.

* Adds a job to build a tar containing the rpc-gating venv and ansible
  roles. This tar is pushed to rpc-repo, and named with the SHA of
  rpc-gating it was built with. The most recent ten tars are kept.
  This job is triggered by pushes to rpc-gating, so new versions
  will automatically built as changes are merged.

* Adds functionality to common.install_ansible to check for a venv tar
  if one exists, it is downloaded, expanded and fixed. Otherwise the
  venv is built locally.

  Jobs won't be blocked between a change merging and the venv tar being
  available on the repo servers, as they will fall back to creating the
  venv locally.