RE-133 Revise artifact build job
Currently the pipelines, if set to replace
existing artifacts, will double-do the apt/git
artifacts which is unnecessary.

The container artifacts for multiple
distributions cannot be done at the same time
due to a race condition when uploading.

This patch therefore takes away the image
and context axes from the jobs and forces the
jobs to run in series or parallel as
appropriate.