[LA-372] influx reporting in gating tooling
Gating could ship this small tool to query influx db based
on env vars. When running a job, we could check for the presence
of the tool, and run it. It automatically returns
the uptime of your job_reference on stdout.

On top of that, we could run the tool with parameters to have
artefacts (YAML/subunit) of the data.

It also allows a parameter for minimum uptime. This is the
success criteria for subunit output.
In case of failure, subunit receives a binary string
as attachment (file_bytes and file_name needs to be given).