RE-133 Pass image param to runonpubcloud
In #370 the runonpubcloud function was modified to take an image
parameter as we need to use different images in parallel. However
the artifact build groovy functions were not updated to supply the image
parameter and will still setting the image env var instead.

Setting the env var causes problems with parallelism as the env
is shared between parallel branches.