RE-133 Reduce file path conflicts
The artifact jobs use the parallel function to run multiple tests in
parallel. These test each require a single use slave, so the slave
allocation and connection functions get run multiple times in parallel
on the same node. This raised multiple issues where the same files were
used by all branches leading to conflicts.

This commit adds random strings to the paths to ensure uniqueness.
Paths changed:
  * venvplaybook vars file path - venvplayook writes all vars to a
    file and supplies that file to ansible-playbook.
  * single use slave inventory - this file stores information about
    the slave.
  * Inventory stash path and name - As the inventory path has changed
    the stash & unstash commands needed to be updated to store the
    correct path. Also the stash name needed to be made unique to
    prevent the stashes overwriting each other.

The allocate_pubcloud playbook is updated to allow inventory_path
to be overridden, and to create the inventory directory if it doesn't
exist.