RE-188 Fix stashing of multi-node inventory
01eac2741173f3016e016f0cbcb31d986ff5c25c modified the process of
stashing the inventory. This change resulted in `connect_deploy_node`
creating a stash that neither used the default name nor passed the
correct name to `ssh_slave.connect`.

This change updates `connect_deploy_node` to use the default.

The setting of the default for `args.inventory` is also fixed so that
the value is correctly assigned.