TURTLES-67 Add gate job for ops fabric chatbot
This is my first attempt at adding a gate job for the
rcbops/ops-fabric-chatbot repo. I want this job to:

- Install any needed dependencies (ops-fabric-chatbot/install-deps.sh)
- Run tests (ops-fabric-chatbot/run-tests.sh)

I would like this job to fail if either of the above commands return
exit code != 0. A timeout of 5 minutes would be good, too, but I haven't
implemented that yet.