RE-199 Improve test for venv availability
Instead of checking the index file, try and download and unpack the
venv tar, if that succeeds unpack it, else create the venv locally.

This is to prevent problems where the tar is listed in the index
but is not fully written/synced, so the job attempts to download it
but either that or unpacking fails.