LA-412 Update influxdb reporting script
This change adds use of `influx-report.sh` to the leapfrog upgrade jobs.
This results in the console log including a breakdown of the downtime
seen by the MaaS plugins during the leapfrog upgrade process.

The Python script `influx.py` has been reworked to allow for a more
detailed report to be produced for leapfrog upgrades. This makes use of
the `.leap` and `.complete` marker files, created at different stages of
the upgrade, to associate the data stored in InfluxDB with the relevant
part of the upgrade. The marker files do not perfectly align with the
upgrades steps but are of sufficient accuracy for use in providing the
breakdown.

If the script is used on a deployment that is not a leapfrog upgrade, it
will still produce overall statistics for the build based solely on the
data in InfluxDB.