TURTLES-67 Run tests in Docker
The scripts run by this job make system wide changes such as installing
apt packages. This means a shared slave cannot be used as multiple
jobs may interact with each other or conflict.

To work around this, tests are executed in a docker container.
Currently an image is built for every job, in future a registry
should be used.