RE-205 Only run tempest_install() on leapfrog
If you now attempt to run the Destroy Slave and Cleanup stages on their
own, the build will attempt to first install tempest.  This may or may
not succeed depending on the state of the AIO.

Since the initial tempest resources are only required for the resource
generation tests run in a leapfrog upgrade, we move this line to within
the leapfrog block.