RE-220 Use fork of openstack-ansible-ops
The current SHA is broken (cobbler26 no longer available), and upstream
has completely rewritten the mnaio build tool which will require a
bunch of RE effort to integrate into rpc-gating.  As a stop gate, we're
going to use a fork of openstack-ansible-ops so we can continue to use
the mnaio jobs until we're able to bump to the head of
openstack-ansible-ops.