LA-436 Fix issues relating to testing instances
The running of `generate-resources.yml` is now managed by rpc-openstack,
the code to perform that function here is removed.

To ensure that Telegraf is properly installed during the upgrade phase
of an upgrade, the appropriate credentials are supplied to the upgrade
script.

To test instance networking a number have vars are set because pinging
using the floating IP address does not work with the default
configuration when using vxlan.