RE-225 Re-order & make RPC-O AIO tests consistent
Re-order the tests to ensure that Tempest runs
before Kibana, then Holland.

The deploy tests all verify that everything is
working for the source branch, so no need to
re-test for upgrades. Also, the MNAIO test
implementation does the more thorough test
regime of before and after.

The tests are now all consistent across all
'actions'. They always run in the same order,
always run for all tests and always use the
same settings. This is simpler to read and
understand what's going on.