RO-2433 Remove support role/holland tests
The support role is being removed from being a part
of RPC-O as it is not an engineering deliverable,
not tested and cannot be reliably included in the
product.

As the support role is also the way that the holland
backup system is deployed, the holland tests are
removed along with it.