TURTLES-67 Add ops fabric analytics service gate
This adds a gate job for the ops-fabric-analytics repo.

I am not sure that I have API access to Jenkins, so I have not tested
this yet.

The goal of this gate is to return failed status if any of the tox (py35
or pep8 tests) fail. JUnit output is nice to have, but I would rather
have a working gate more quickly if possible.

The one part I'm not sure of is if tox exists in the container at
the time the tests are run.